+++ src/model/EnemySystem.h
@@ -13,7 +13,7 @@ namespace isaac::model {
 
 enum class MoveStrategy { Chase, Wander, KeepDistance, Dash, Stationary };
 enum class AttackStrategy { Contact, AimedShot, RadialShot };
-enum class DropStrategy { None, Coin, Bomb, Key };
+enum class DropStrategy { None, Breakfast };
 
 struct EnemyDefinition {
   std::string_view id;